Move boundary test into separate test file.
